cleanup query cache structure
