+++ grakn-graql/src/main/java/ai/grakn/graql/internal/reasoner/ResolutionIterator.java
@@ -20,7 +20,7 @@ package ai.grakn.graql.internal.reasoner;
 
 import ai.grakn.graql.admin.Answer;
 import ai.grakn.graql.internal.query.QueryAnswer;
-import ai.grakn.graql.internal.reasoner.cache.QueryCache;
+import ai.grakn.graql.internal.reasoner.cache.SimpleQueryCache;
 import ai.grakn.graql.internal.reasoner.iterator.ReasonerQueryIterator;
 import ai.grakn.graql.internal.reasoner.query.ReasonerAtomicQuery;
 import ai.grakn.graql.internal.reasoner.query.ReasonerQueryImpl;
@@ -49,7 +49,7 @@ public class ResolutionIterator extends ReasonerQueryIterator {
     private final ReasonerQueryImpl query;
     private final Set<Answer> answers = new HashSet<>();
 
-    private final QueryCache<ReasonerAtomicQuery> cache = new QueryCache<>();
+    private final SimpleQueryCache<ReasonerAtomicQuery> cache = new SimpleQueryCache<>();
     private final Stack<ResolutionState> states = new Stack<>();
 
     private Answer nextAnswer = null;
