+++ grakn-graql/src/main/java/ai/grakn/graql/internal/reasoner/state/AtomicStateProducer.java
@@ -21,7 +21,7 @@ package ai.grakn.graql.internal.reasoner.state;
 import ai.grakn.graql.admin.Answer;
 import ai.grakn.graql.admin.Unifier;
 import ai.grakn.graql.internal.reasoner.atom.predicate.NeqPredicate;
-import ai.grakn.graql.internal.reasoner.cache.QueryCache;
+import ai.grakn.graql.internal.reasoner.cache.SimpleQueryCache;
 import ai.grakn.graql.internal.reasoner.query.ReasonerAtomicQuery;
 import com.google.common.collect.Iterators;
 import java.util.Iterator;
@@ -43,7 +43,7 @@ public class AtomicStateProducer extends QueryStateBase {
 
     private final Iterator<ResolutionState> subGoalIterator;
 
-    public AtomicStateProducer(ReasonerAtomicQuery query, Answer sub, Unifier u, QueryStateBase parent, Set<ReasonerAtomicQuery> subGoals, QueryCache<ReasonerAtomicQuery> cache) {
+    public AtomicStateProducer(ReasonerAtomicQuery query, Answer sub, Unifier u, QueryStateBase parent, Set<ReasonerAtomicQuery> subGoals, SimpleQueryCache<ReasonerAtomicQuery> cache) {
         super(sub, u, parent, subGoals, cache);
 
         if(query.getAtom().getSchemaConcept() == null){
