+++ grakn-graql/src/main/java/ai/grakn/graql/internal/reasoner/state/ConjunctiveState.java
@@ -22,7 +22,7 @@ import ai.grakn.graql.admin.Answer;
 import ai.grakn.graql.admin.MultiUnifier;
 import ai.grakn.graql.admin.Unifier;
 import ai.grakn.graql.internal.reasoner.MultiUnifierImpl;
-import ai.grakn.graql.internal.reasoner.cache.QueryCache;
+import ai.grakn.graql.internal.reasoner.cache.SimpleQueryCache;
 import ai.grakn.graql.internal.reasoner.query.ReasonerAtomicQuery;
 import ai.grakn.graql.internal.reasoner.query.ReasonerQueries;
 import ai.grakn.graql.internal.reasoner.query.ReasonerQueryImpl;
@@ -45,7 +45,7 @@ public class ConjunctiveState extends QueryState<ReasonerQueryImpl> {
                             Unifier u,
                             QueryStateBase parent,
                             Set<ReasonerAtomicQuery> visitedSubGoals,
-                            QueryCache<ReasonerAtomicQuery> cache) {
+                            SimpleQueryCache<ReasonerAtomicQuery> cache) {
         super(ReasonerQueries.create(q, sub), sub, u, MultiUnifierImpl::new, parent, visitedSubGoals, cache);
     }
